put all ModelChecking trackers into ModelCheckingStrategy.kt
